move featuredprojects fetch into component
